+++ tests/setup.ts
@@ -8,5 +8,6 @@ Object.assign(process.env, {
   GITHUB_REPOSITORY: 'JasonEtco/waddup',
   GITHUB_ACTION: 'create-a-github-issue',
   GITHUB_EVENT_PATH: path.join(__dirname, 'fixtures', 'event.json'),
-  GITHUB_WORKSPACE: path.join(__dirname, 'fixtures')
+  GITHUB_WORKSPACE: path.join(__dirname, 'fixtures'),
+  GITHUB_TOKEN: 'test-token'
 })