+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### 3.3.0 (Next)
 
 * Your contribution here.
-* [#168](https://github.com/dblock/create-a-github-issue/pull/XXX): Upgraded @actions/core to 3.0 with ESM support - [@dblock](https://github.com/dblock).
+* [#168](https://github.com/dblock/create-a-github-issue/pull/168): Upgraded @actions/core to 3.0 with ESM support - [@dblock](https://github.com/dblock).
+* [#169](https://github.com/dblock/create-a-github-issue/pull/169): Removed deprecated actions-toolkit, use @actions/core and @actions/github directly - [@dblock](https://github.com/dblock).
 
 ### 3.2.0 (2024/04/25)